--- ExpressionEvaluator/Io.JoeMoceri.ExpressionEvaluator/LanguageTemplate/LanguageTemplateType.cs
@@ -1,8 +1,0 @@
-﻿namespace ExpressionEvaluator
-{
-    // create new language templates here
-    public enum LanguageTemplateType
-    {
-        EEExpressions
-    }
-}
